requirements.txt
smolagents>=0.6.0
akshare>=1.12.0
markdown2>=2.4.0
weasyprint>=60.0
pyyaml>=6.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
openai>=1.0.0
yfinance>=0.2.38
